backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyserial==3.5
pydantic==2.5.0
pydantic-settings==2.1.0

